+++ coredns/datadog_checks/coredns/__init__.py
@@ -1,0 +1,11 @@
+# (C) Datadog, Inc. 2018
+# All rights reserved
+# Licensed under a 3-clause BSD style license (see LICENSE)
+
+from .coredns import CoreDNSCheck
+from .__about__ import __version__
+
+__all__ = [
+    '__version__',
+    'CoreDNSCheck'
+]