+++ datadog_checks_dev/datadog_checks/dev/tooling/utils.py
@@ -7,8 +7,9 @@ import re
 from ast import literal_eval
 
 import requests
+import semver
 
-from .constants import get_root
+from .constants import VERSION_BUMP, get_root
 from ..utils import file_exists, read_file
 
 # match something like `(#1234)` and return `1234` in a group
@@ -68,7 +69,7 @@ def string_to_toml_type(s):
 
 
 def get_version_file(check_name):
-    if check_name in ('datadog_checks_base', 'datadog_checks_test_helper'):
+    if check_name == 'datadog_checks_base':
         return os.path.join(get_root(), check_name, 'datadog_checks', '__about__.py')
     elif check_name == 'datadog_checks_dev':
         return os.path.join(get_root(), check_name, 'datadog_checks', 'dev', '__about__.py')
@@ -109,3 +110,16 @@ def load_manifest(check_name):
     if file_exists(manifest_path):
         return json.loads(read_file(manifest_path))
     return {}
+
+
+def get_bump_function(changelog_types):
+    minor_bump = False
+
+    for changelog_type in changelog_types:
+        bump_function = VERSION_BUMP.get(changelog_type)
+        if bump_function is semver.bump_major:
+            return bump_function
+        elif bump_function is semver.bump_minor:
+            minor_bump = True
+
+    return semver.bump_minor if minor_bump else semver.bump_patch