+++ agent_metrics/CHANGELOG.md
@@ -1,5 +1,10 @@
 # CHANGELOG - agent_metrics
 
+## 1.1.1 / 2018-09-04
+
+* [Fixed] Moves Agent Metrics to Pytest. See [#2059](https://github.com/DataDog/integrations-core/pull/2059).
+* [Fixed] Add data files to the wheel package. See [#1727](https://github.com/DataDog/integrations-core/pull/1727).
+
 ## 1.1.0 / 2018-03-23
 
 * [FEATURE] Add custom tag support.