+++ dns_check/CHANGELOG.md
@@ -1,5 +1,9 @@
 # CHANGELOG - dns_check
 
+## 1.1.2 / 2018-09-04
+
+* [Fixed] Add data files to the wheel package. See [#1727](https://github.com/DataDog/integrations-core/pull/1727).
+
 ## 1.1.1 / 2018-06-13
 
 * [Fixed] Fix parsing of requirements file so that `dnspython` is listed as a req. See [#1603](https://github.com/DataDog/integrations-core/pull/1603).