+++ kong/CHANGELOG.md
@@ -1,5 +1,9 @@
 # CHANGELOG - kong
 
+## 1.2.1 / 2018-09-04
+
+* [Fixed] Add data files to the wheel package. See [#1727](https://github.com/DataDog/integrations-core/pull/1727).
+
 ## 1.2.0 / 2018-05-11
 
 * [FEATURE] Add `ssl_validation` settings to disable SSL Cert Verification